requirements.txt
flask
chromadb
google-genai
python-dotenv
pandas
